add multistep mission pairings UI
